BERT 출력 FP32 캐스팅 + 벤치마크 출력 gitignore
bert_feature.py: BERT가 FP16/Q8일 때 출력을 FP32로 캐스팅하여
Synthesizer dtype 불일치 방지.

+++ style_bert_vits2/nlp/japanese/bert_feature.py
@@ -51,7 +51,7 @@ def extract_bert_feature(
         for i in inputs:
             inputs[i] = inputs[i].to(device)  # type: ignore
         res = model(**inputs, output_hidden_states=True)
-        res = torch.cat(res["hidden_states"][-3:-2], -1)[0].cpu()
+        res = torch.cat(res["hidden_states"][-3:-2], -1)[0].cpu().float()
         if assist_text:
             style_inputs = tokenizer(assist_text, return_tensors="pt")
             for i in style_inputs:
